+++ CHANGELOG.md
@@ -9,6 +9,10 @@ use patch releases for compatibility fixes instead.
 
 ## Unreleased
 
+### Fixed
+
+- Added support for uploading empty files to datasets.
+
 ## [v1.18.3](https://github.com/allenai/beaker-py/releases/tag/v1.18.3) - 2023-04-19
 
 ### Added